changes:
1.  remove dimension with size 1.
2. combine dimensions that's still in order after reorder
3. add a 3D tiled reshape.

many reshape can be use 2D transpose with cublas or 3D tiled reshape with cuda now, both are 4x faster.
for example: {1, 1, 1024, 256} -> {256, 1024, 1, 1} with input order {3, 2, 0, 1}
      become: {1024, 256} -> {256, 1024} with input order {1, 0}
which can use cublas matrix transpose.

for example: {256, 256, 3, 3} -> {3, 3, 256, 256} with input order {3, 2, 0, 1}
      become: {256, 256, 9} -> {9, 256, 256} with input order {2, 1, 0}
which can use 3D tiled reshape.

TODO: nd tiled reshape

```
Benchmarking ../ngraph-models/models/tensorflow/inference/resnet50/tf_function_ngraph_cluster_rn50_bs1.json, GPU backend, 10 iterations.
compile time: 7,043ms
11.5ms per iteration

---- Aggregate times per op type ----
Convolution     7,766us
Reshape         2,968us
BatchNorm         497us
Relu              334us
Add               202us
MaxPool            59us
Dot                34us
Sum                13us
Pad                10us
Divide              5us
Broadcast           5us
Result              1us
```

after:
```
Benchmarking ../ngraph-models/models/tensorflow/inference/resnet50/tf_function_ngraph_cluster_rn50_bs1.json, GPU backend, 10 iterations.
compile time: 7,050ms
9.6ms per iteration

---- Aggregate times per op type ----
Convolution     7,737us
Reshape         1,096us
BatchNorm         505us
Relu              328us
Add               203us
MaxPool            59us
Dot                34us
Sum                13us
Pad                10us
Divide              5us
Broadcast           4us
Result              2us
```